backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-socketio==5.11.1
python-multipart==0.0.9
